opendal_service_swift

Struct Swift

pub struct Swift { /* private fields */ }

OpenStack Swift’s REST API support. For more information about swift-compatible services, refer to Compatible Services.

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::SwiftConfig for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Examples

Via Builder

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_swift::Swift;

#[tokio::main]
async fn main() -> Result<()> {
    // Create Swift backend builder
    let mut builder = Swift::default()
        // Set the root for swift, all operations will happen under this root
        .root("/path/to/dir")
        // set the endpoint of Swift backend
        .endpoint("https://openstack-controller.example.com:8080/v1/account")
        // set the container name of Swift workspace
        .container("container")
        // set the auth token for builder
        .token("token");

    let op: Operator = Operator::new(builder)?;

    Ok(())
}

Compatible Services

OpenStack Swift

OpenStack Swift is the default implementations of swift services.

To connect to OpenStack Swift, we need to set:

endpoint: The endpoint of OpenStack Swift, for example: http://127.0.0.1:8080/v1/AUTH_test.

container: The name of OpenStack Swift container.

token: OpenStack Swift container personal access token.

builder.endpoint("http://127.0.0.1:8080/v1/AUTH_test");
builder.container("container");
builder.token("token");

endpoint is the full URL that serves as the access point to all containers under an OpenStack Swift account. It represents the entry point for accessing the resources of the account. Alongside endpoint, token is used as a credential to verify the user’s identity and authorize access to the relevant resources. Both endpoint and token can be obtained through OpenStack Swift authentication service.

endpoint consists of server address and port, API version, authenticated account ID. For instance, it might appear as follows:

http://127.0.0.1:8080/v1/AUTH_test.

http://192.168.66.88:8080/swift/v1.

https://openstack-controller.example.com:8080/v1/account.

Please note that the exact format of endpoint may vary depending on the deployment configuration and version of swift services. Users can refer to the specific services documentation for the correct endpoint format and authentication method.

For more information, refer:

OpenStack Swift API.

OpenStack Swift Authentication.

Ceph Rados Gateway

Ceph Rados Gateway supports a RESTful API that is compatible with the basic data access model of OpenStack Swift API.

To connect to Ceph Rados Gateway, we need to set:

endpoint: The endpoint of swift services, for example: http://127.0.0.1:8080/swift/v1.

container: The name of swift container.

token: swift container personal access token.

builder.endpoint("http://127.0.0.1:8080/swift/v1");
builder.container("container");
builder.token("token");

For more information, refer:

Ceph Rados Gateway Swift API.

Ceph Rados Gateway Swift Authentication.

Implementations

impl SwiftBuilder

pub fn endpoint(self, endpoint: &str) -> Self

Set the remote address of this backend

Endpoints should be full uri, e.g.

http://127.0.0.1:8080/v1/AUTH_test

http://192.168.66.88:8080/swift/v1

https://openstack-controller.example.com:8080/v1/ccount

If user inputs endpoint without scheme, we will prepend https:// to it.

pub fn container(self, container: &str) -> Self

Set container of this backend.

All operations will happen under this container. It is required. e.g. snapshots

pub fn root(self, root: &str) -> Self

Set root of this backend.

All operations will happen under this root.

pub fn token(self, token: &str) -> Self

Set the token of this backend.

Default to empty string.

pub fn temp_url_key(self, key: &str) -> Self

Set the TempURL key for generating presigned URLs.

This should match the X-Account-Meta-Temp-URL-Key or X-Container-Meta-Temp-URL-Key value configured on the Swift account or container.

pub fn temp_url_hash_algorithm(self, algo: &str) -> Self

Set the hash algorithm for TempURL signing.

Supported values: sha1, sha256, sha512. Defaults to sha256. The cluster must have the chosen algorithm in its tempurl.allowed_digests (check GET /info).